==> 02_Feature_Engineering.ipynb <==
#notebook for engineer features using game results and stats
#import necessary packages
import pandas as pd
import numpy as np

#import Data
PL19_20_fixtures = pd.read_csv('Data/2019-20/Fixtures/PL19_20_fixtures.csv')
PL20_21_fixtures = pd.read_csv('Data/2020-21/Fixtures/PL20_21_fixtures.csv')
PL21_22_fixtures = pd.read_csv('Data/2021-22/Fixtures/PL21_22_fixtures.csv')
PL19_20_fixtures['game_home_ID'] = PL19_20_fixtures['HomeTeam'] + PL19_20_fixtures['Date'].astype(str)
PL19_20_fixtures['game_away_ID'] = PL19_20_fixtures['AwayTeam'] + PL19_20_fixtures['Date'].astype(str)
PL20_21_fixtures['game_home_ID'] = PL20_21_fixtures['HomeTeam'] + PL20_21_fixtures['Date'].astype(str)
PL20_21_fixtures['game_away_ID'] = PL20_21_fixtures['AwayTeam'] + PL20_21_fixtures['Date'].astype(str)
PL21_22_fixtures['game_home_ID'] = PL21_22_fixtures['HomeTeam'] + PL21_22_fixtures['Date'].astype(str)
PL21_22_fixtures['game_away_ID'] = PL21_22_fixtures['AwayTeam'] + PL21_22_fixtures['Date'].astype(str)

#generate gameweeks dependening on the order of the games
def generate_gameweeks(df):
    teams = df.HomeTeam.unique().tolist()
    df['Wk_h'] = 0
    df['Wk_a'] = 0
    for team in teams:
        count = 1
        for index, row in df.iterrows():
            if row['HomeTeam'] == team:
                df.at[index, 'Wk_h'] = count
                count = count + 1
            elif row['AwayTeam'] == team:
                df.at[index, 'Wk_a'] = count
                count = count + 1
    return df
PL19_20_fixtures = generate_gameweeks(PL19_20_fixtures)
PL20_21_fixtures = generate_gameweeks(PL20_21_fixtures)
PL21_22_fixtures = generate_gameweeks(PL21_22_fixtures)

PL21_22_fixtures

#functions for feature Engineering

#get points for one game
def get_points(team, wk, df):
    for index, row in df.iterrows():
        if row['Wk_h'] == wk and row['HomeTeam'] == team:
            if row['FTR'] == 'H':
                return 3
            if row['FTR'] == 'D':
                return 1
            else:
                return 0
        if row['Wk_a'] == wk and row['AwayTeam'] == team:
            if row['FTR'] == 'A':
                return 3
            if row['FTR'] == 'D':
                return 1
            else:
                return 0

#get values for statistics (shots, corners, goals)
def get_stats(team, wk, stat, df):
    for index, row in df.iterrows():
        #print(row[['HomeTeam','Wk']])
        if row['Wk_h'] == wk and row['HomeTeam'] == team:
            if stat == 'goals':
                return row['FTHG']
            elif stat == 'shots':
                return row['HST']
            elif stat == 'corners':
                return row['HC']
        elif row['Wk_a'] == wk and row['AwayTeam'] == team:
            if stat == 'goals':
                return row['FTAG']
            elif stat == 'shots':
                return row['AST']
            elif stat == 'corners':
                return row['AC']
            return row[stat]

#sum points over a season
def get_points_season(team, fr, to, df):
    test = []
    out = 0
    df2 = 0
    if fr == to: return 0
    for i in range(fr, to):
        out = out + get_points(team, i, df)
        test.append(out)
        df3 = pd.Series(test)
        df2 = df3.rolling(len(df3)).mean()
        df2 = df2[len(df3) - 1]
    return df2

#sum stats values over a season
def get_stats_season(team, fr, to, category, df):
    test = []
    out = 0
    df2 = 0
    if fr == to: return 0
    for i in range(fr, to):
        out = out + get_stats(team, i, category, df)
        test.append(out)
        df3 = pd.Series(test)
        df2 = df3.rolling(len(df3)).mean()
        df2 = df2[len(df3) - 1]
        #df2 = df.ewm(alpha=0.4).mean()[len(df) - 1]
    return df2

#get ewm over a season
def get_ewm_season(team, fr, to, df):
    test = []
    out = 0
    if fr == to: return 0
    for i in range(fr, to):
        out = out + get_points(team, i, df)
        test.append(out)
        df3 = pd.Series(test)
        df2 = df3.ewm(alpha=0.9).mean()[len(df3) - 1]
    return df2

#append form to df
def append_form(timespan, df):
    formH = []
    formA = []
    for index, row in df.iterrows():
        wk_h = row['Wk_h']
        wk_a = row['Wk_a']
        teamH = row['HomeTeam']
        teamA = row['AwayTeam']
        if int(wk_h) - timespan < 1:
            pointsH = get_points_season(teamH, 1, int(wk_h) - 1, df)
        else:
            pointsH = get_points_season(teamH, int(wk_h) - timespan, int(wk_h) - 1, df=df)
        if int(wk_a) - timespan < 1:
            pointsA = get_points_season(teamA, 1, int(wk_a) - 1, df)
        else:
            pointsA = get_points_season(teamA, int(wk_a) - timespan, int(wk_a) - 1, df=df)
        formH.append(pointsH)
        formA.append(pointsA)
    df['formH'] = formH
    df['formA'] = formA
    return df

#append ewm to df
def append_ewm(timespan, df):
    formH = []
    formA = []
    for index, row in df.iterrows():
        wk_h = row['Wk_h']
        wk_a = row['Wk_a']
        teamH = row['HomeTeam']
        teamA = row['AwayTeam']
        if int(wk_h) - timespan < 1:
            pointsH = get_ewm_season(teamH, 1, int(wk_h), df)
        else:
            pointsH = get_ewm_season(teamH, int(wk_h) - timespan, int(wk_h) - 1, df)
        if int(wk_a) - timespan < 1:
            pointsA = get_ewm_season(teamA, 1, int(wk_a), df)
        else:
            pointsA = get_ewm_season(teamA, int(wk_a) - timespan, int(wk_a) - 1, df)
        formH.append(pointsH)
        formA.append(pointsA)
    df['ewmH'] = formH
    df['ewmA'] = formA
    return df

#append stats to df
def append_stats(timespan, df):
    goals_h = []
    goals_A = []
    shots_H = []
    shots_A = []
    corners_H = []
    corners_A = []

    for index, row in df.iterrows():
        wk_h = row['Wk_h']
        wk_a = row['Wk_a']
        teamH = row['HomeTeam']
        teamA = row['AwayTeam']
        if int(wk_h) - timespan < 1:
            goalsH = get_stats_season(teamH, 1, int(wk_h) - 1, 'goals', df)
            shotsH = get_stats_season(teamH, 1, int(wk_h) - 1, 'shots', df)
            cornersH = get_stats_season(teamH, 1, int(wk_h) - 1, 'corners', df)
        else:
            goalsH = get_stats_season(teamH, int(wk_h) - timespan, int(wk_h) - 1, 'goals', df)
            shotsH = get_stats_season(teamH, int(wk_h) - timespan, int(wk_h) - 1, 'shots', df)
            cornersH = get_stats_season(teamH, int(wk_h) - timespan, int(wk_h) - 1, 'corners', df)

        if int(wk_a) - timespan < 1:
            goalsA = get_stats_season(teamA, 1, int(wk_a) - 1, 'goals', df)
            shotsA = get_stats_season(teamA, 1, int(wk_a) - 1, 'shots', df)
            cornersA = get_stats_season(teamA, 1, int(wk_a) - 1, 'corners', df)
        else:
            goalsA = get_stats_season(teamA, int(wk_a) - timespan, int(wk_a) - 1, 'goals', df)
            shotsA = get_stats_season(teamA, int(wk_a) - timespan, int(wk_a) - 1, 'shots', df)
            cornersA = get_stats_season(teamA, int(wk_a) - timespan, int(wk_a) - 1, 'corners', df)

        goals_h.append(goalsH)
        goals_A.append(goalsA)
        shots_H.append(shotsH)
        shots_A.append(shotsA)
        corners_H.append(cornersH)
        corners_A.append(cornersA)
    df['goalsH'] = goals_h
    df['goalsA'] = goals_A
    df['shotsH'] = shots_H
    df['shotsA'] = shots_A
    df['cornersH'] = corners_H
    df['cornersA'] = corners_A
    return df


#Calculate Elo Rating
def expected_result(loc, aw):
    dr = loc - aw
    we = (1 / (10 ** (-dr / 400) + 1))
    return [np.round(we, 3), 1 - np.round(we, 3)]

def actual_result(loc, aw):
    if loc < aw:
        wa = 1
        wl = 0
    elif loc > aw:
        wa = 0
        wl = 1
    elif loc == aw:
        wa = 0.5
        wl = 0.5
    return [wl, wa]


def calculate_elo(elo_l, elo_v, local_goals, away_goals, k_value):
    k = k_value
    wl, wv = actual_result(local_goals, away_goals)
    wel, wev = expected_result(elo_l, elo_v)

    elo_ln = elo_l + k * (wl - wel)
    elo_vn = elo_v + k * (wv - wev)

    return elo_ln, elo_vn

def calculate_elo_season(df):
    current_elo = {}
    for idx, row in df.iterrows():

        local = row['HomeTeam']
        away = row['AwayTeam']
        local_goals = row['FTHG']
        away_goals = row['FTAG']
        #tournament = row['tournament']

        if local not in current_elo.keys():
            current_elo[local] = 1500

        if away not in current_elo.keys():
            current_elo[away] = 1500

        elo_l = current_elo[local]
        elo_v = current_elo[away]
        elo_ln, elo_vn = calculate_elo(elo_l, elo_v, local_goals, away_goals, 30)

        current_elo[local] = elo_ln
        current_elo[away] = elo_vn

        df.loc[idx, 'Elo_h_after'] = elo_ln
        df.loc[idx, 'Elo_a_after'] = elo_vn
        df.loc[idx, 'Elo_h_before'] = elo_l
        df.loc[idx, 'Elo_a_before'] = elo_v
        return df
calculate_elo_season(PL21_22_fixtures)

PL19_20_fixtures = calculate_elo_season(PL19_20_fixtures)
PL21_22_fixtures = calculate_elo_season(PL21_22_fixtures)

#add features to table
PL19_20_fixtures = append_stats(7, PL19_20_fixtures)
PL19_20_fixtures = append_ewm(40, PL19_20_fixtures)
PL19_20_fixtures = append_form(7, PL19_20_fixtures)
PL19_20_fixtures = calculate_elo_season(PL19_20_fixtures)
PL20_21_fixtures = append_stats(7, PL20_21_fixtures)
PL20_21_fixtures = append_ewm(40, PL20_21_fixtures)
PL20_21_fixtures = append_form(7, PL20_21_fixtures)
PL20_21_fixtures = calculate_elo_season(PL20_21_fixtures)
PL21_22_fixtures = append_stats(7, PL21_22_fixtures)
PL21_22_fixtures = append_ewm(40, PL21_22_fixtures)
PL21_22_fixtures = append_form(7, PL21_22_fixtures)
PL21_22_fixtures = calculate_elo_season(PL21_22_fixtures)


#calculate Differences
def calculate_differences(df):
    df['goalsDifference'] = df['goalsH'] - df['goalsA']
    df['shotsDifference'] = df['shotsH'] - df['shotsA']
    df['cornersDifference'] = df['cornersH'] - df['cornersA']
    df['ewmDifference'] = df['ewmH'] - df['ewmA']
    df['elo_difference'] = df['Elo_h_before'] - df['Elo_a_before']
    df['form_difference'] = df['formH'] - df['formA']
    return df
PL19_20_fixtures = calculate_differences(PL19_20_fixtures)
PL20_21_fixtures = calculate_differences(PL20_21_fixtures)
PL21_22_fixtures = calculate_differences(PL21_22_fixtures)

PL20_21_fixtures

#export features as csv
#PL19_20_fixtures.to_csv('Data/2019-20/Fixtures/PL19_20_features.csv')
#PL20_21_fixtures.to_csv('Data/2020-21/Fixtures/PL20_21_features.csv')
#PL21_22_fixtures.to_csv('Data/2021-22/Fixtures/PL21_22_features.csv')
