==> 03_Webscraping_Fbref.ipynb <==
#notebook for scrape all necessary player stats nad lineups from web
#import necessary packages
import requests
from bs4 import BeautifulSoup
import pandas as pd
import numpy as np
import re
import csv
import time

#method for scraping links for all games of a EPL season
def get_links(url):
    res = requests.get(url)
    ## The next two lines get around the issue with comments breaking the parsing.
    comm = re.compile("<!--|-->")
    soup = BeautifulSoup(comm.sub("", res.text), 'lxml')
    all_tables = soup.findAll("tbody")
    test = soup.find(id='div_sched_2021-2022_9_1')
    x = "score"
    test = test.findAll("td", {"data-stat": x})
    link_list = []
    for i in test:
        i = i.find("a")
        if i is not None:
            link = i['href']
            link_list.append(link)
    return link_list

#methods for scraping

#stats matchday
statsm = ["player", "shirtnumber", "nationality", "position", "age", "minutes", "goals", "assists", "touches", "sca",
          "gca", 'pens_made', 'shots_total', 'shots_on_target', 'pressures', 'tackles', 'interceptions', 'blocks', 'xg',
          'xa', 'passes', 'passes_pct', 'progressive_passes', 'progressive_carries', 'dribbles_completed', 'dribbles']

#stats keeper matchday
statskeepers = ['player', 'minutes', 'age', 'gk_shots_on_target_against', 'gk_goals_against', 'gk_saves', 'gk_save_pct',
                'gk_psxg', 'gk_passes_completed_launched', 'gk_passes', 'gk_passes_throws', 'gk_pct_passes_launched',
                'gk_def_actions_outside_pen_area']

keepers = ["player", "nationality", "position", "squad", "age", "birth_year", "games_gk", "games_starts_gk",
           "minutes_gk", "goals_against_gk", "goals_against_per90_gk", "shots_on_target_against", "saves", "save_pct",
           "wins_gk", "draws_gk", "losses_gk", "clean_sheets", "clean_sheets_pct", "pens_att_gk", "pens_allowed",
           "pens_saved", "pens_missed_gk"]
keepers3 = ["players_used", "games_gk", "games_starts_gk", "minutes_gk", "goals_against_gk", "goals_against_per90_gk",
            "shots_on_target_against", "saves", "save_pct", "wins_gk", "draws_gk", "losses_gk", "clean_sheets",
            "clean_sheets_pct", "pens_att_gk", "pens_allowed", "pens_saved", "pens_missed_gk"]




#Functions to get the data in a dataframe using BeautifulSoup

#access tables from fbref
def get_tables_m(url):
    res = requests.get(url)
    ## The next two lines get around the issue with comments breaking the parsing.
    comm = re.compile("<!--|-->")
    soup = BeautifulSoup(comm.sub("", res.text), 'lxml')
    all_tables = soup.findAll("tbody")
    matchweek = soup.find(id='content').find("div").contents[1]
    date = soup.find("div", {"class": "scorebox_meta"}).find("div").contents[0].contents[0]
    date = date['href']
    date = date[12:22]

    team_H = soup.findAll("strong")[2].find("a").contents[0]
    if soup.findAll("strong")[5].find("a") is None:
        team_A = soup.findAll("strong")[4].find("a").contents[0]
    else:
        team_A = soup.findAll("strong")[5].find("a").contents[0]

    team1_table = all_tables[0]
    team2_table = all_tables[7]
    gk1_table = all_tables[6]
    gk2_table = all_tables[13]

    return team1_table, team2_table, gk1_table, gk2_table, matchweek, date, team_H, team_A


#extract information from tables
def get_frame(features, player_table):
    pre_df_player = dict()
    features_wanted_player = features
    rows_player = player_table.find_all('tr')
    rows_player
    #print(rows_player)
    for row in rows_player:
        #print(row)
        if (row.find('th', {"scope": "row"}) != None):

            for f in features_wanted_player:
                #print(f)
                cell = row.find("td", {"data-stat": f})
                #print(cell)
                if cell is None:
                    text = row.find('th', {"scope": "row"}).find('a').contents[0]
                    if f in pre_df_player:
                        pre_df_player[f].append(text)
                    else:
                        pre_df_player[f] = [text]
                #print(cell)
                else:
                    a = cell.text.strip().encode()
                    text = a.decode("utf-8")
                    #print(text)
                    if (text == ''):
                        text = '0'
                    if ((f != 'player') & (f != 'nationality') & (f != 'position') & (f != 'squad') & (f != 'age') & (
                            f != 'birth_year')):
                        text = float(text.replace(',', ''))
                    if f in pre_df_player:
                        pre_df_player[f].append(text)
                    else:
                        pre_df_player[f] = [text]
    df_player = pd.DataFrame.from_dict(pre_df_player)
    return df_player

#get information in df
def frame_for_category(category, top, end, features):
    url = (top + category + end)
    print(url)
    player_table_H, player_table_A, gk_table_H, gk_table_A, matchweek, date, team_H, team_A = get_tables_m(url)
    #print(player_table)
    df_player_H = get_frame(features, player_table_H)
    df_player_H['Team'] = team_H
    df_player_A = get_frame(features, player_table_A)
    df_player_A['Team'] = team_A
    df_player = pd.concat([df_player_H, df_player_A])
    df_player['matchweek'] = matchweek
    df_player['date'] = date
    df_gk_H = get_frame(statskeepers, gk_table_H)
    df_gk_H['Team'] = team_H
    df_gk_A = get_frame(statskeepers, gk_table_A)
    df_gk_A['Team'] = team_A
    df_gk = pd.concat([df_gk_H, df_gk_A])
    df_gk['matchweek'] = matchweek
    df_gk['date'] = date
    return df_player, df_gk


#get matchday data for one season
def get_matchday_data(top, end, linklist):
    pl, gk = frame_for_category('/en/matches/3adf2aa7', top, end, statsm)

    for i in linklist:
        tmppl, tmpgk = frame_for_category(i, top, end, statsm)
        pl = pd.concat([pl, tmppl], axis=0)
        gk = pd.concat([gk, tmpgk], axis=0)
        time.sleep(5)

    return pl, gk


#get lineups for each match

#access tables from fbref
def get_lineups(url):
    pre_df_player = dict()
    res = requests.get(url)
    ## The next two lines get around the issue with comments breaking the parsing.
    comm = re.compile("<!--|-->")
    soup = BeautifulSoup(comm.sub("", res.text), 'lxml')
    all_tables = soup.findAll("tbody")
    date = soup.find("div", {"class": "scorebox_meta"}).find("div").contents[0].contents[0]
    date = date['href']
    date = date[12:22]
    matchweek = soup.find(id='content').find("div").contents[1]
    lineup = soup.find_all("div", {"class": "lineup"})
    lineup_H = lineup[0]
    lineup_A = lineup[1]

    rows_lineup_H = lineup_H.find_all('tr')
    rows_lineup_A = lineup_A.find_all('tr')
    lineup_total = [rows_lineup_H, rows_lineup_A]
    #print(rows_lineup)
    i = 0


    for rows_lineup in lineup_total:
        i = i+1
        for row in rows_lineup:

            if row.find('a') != None:
                #print(row.find('a').contents[0])
                cell = row.find('a').contents[0]
                a = cell.text.strip().encode()
                text = a.decode("utf-8")

                if 'player' in pre_df_player:
                    pre_df_player['player'].append(text)
                    #pre_df_player['count'].append(3)
                else:
                    pre_df_player['player'] = [text]
                    #pre_df_player['count'] = i

        if i == 1:
            df_player_H = pd.DataFrame.from_dict(pre_df_player)
        else: df_player_A = pd.DataFrame.from_dict(pre_df_player)
        pre_df_player = dict()



    team_H = soup.findAll("strong")[2].find("a").contents[0]
    if soup.findAll("strong")[5].find("a") is None:
        team_A = soup.findAll("strong")[4].find("a").contents[0]
    else:
        team_A = soup.findAll("strong")[5].find("a").contents[0]

    #gk1_table.append(gk2_table)
    #team1_table.append(team2_table)

    return df_player_H, df_player_A, team_H, team_A, matchweek, date

#extract information from tables
def frame_for_category_lineups (category, top, end):
    url = (top + category + end)
    print(url)
    player_H, player_A, teamH, teamA, matchweek, date,= get_lineups(url)

    player_H['team'] = teamH
    player_A['team'] = teamA
    df_player = pd.concat([player_H, player_A])
    df_player['matchweek'] = matchweek
    df_player['date'] = date
    df_player['start11'] = np.where(df_player.index< 11, True, False)

    return df_player

#get lineups for one season
def get_matchday_data_lineups(top, end, linklist):
    pl = frame_for_category_lineups('/en/matches/3adf2aa7', top, end)
    #pl1, gk1 = frame_for_category('/en/matches/3adf2aa7', top, end, statsm)

    for i in linklist:
        tmppl = frame_for_category_lineups(i, top, end)
        pl = pd.concat([pl, tmppl], axis=0)
        time.sleep(5)

    return pl


#get links for one season
link_list = get_links('https://fbref.com/en/comps/9/2021-2022/schedule/2021-2022-Premier-League-Scores-and-Fixtures')

# split list in 4 parts to not bring down website (website crashes sometimes, with split in 4 parts not the complete scraping process has to be done again)
list1 = []
list2 = []
list3 = []
list4 = []
count = 0

for i in link_list:
    if count < 100:
        list1.append(i)
    elif count < 200:
        list2.append(i)
    elif count < 300:
        list3.append(i)
    else:
        list4.append(i)
    count = count + 1

#scrape data and store it in df's for players and goalkeepers
df_1, gk_1 = get_matchday_data('https://fbref.com/', '', list1)
df_2, gk_2 = get_matchday_data('https://fbref.com/', '', list2)
df_3, gk_3 = get_matchday_data('https://fbref.com/', '', list3)
df_4, gk_4 = get_matchday_data('https://fbref.com/', '', list4)

#scrape lineups
l_1 = get_matchday_data_lineups('https://fbref.com/', '', list1)
l_2 = get_matchday_data_lineups('https://fbref.com/', '', list2)
l_3 = get_matchday_data_lineups('https://fbref.com/', '', list3)
l_4 = get_matchday_data_lineups('https://fbref.com/', '', list4)

#concatenate all files
def concatenate_files(f1, f2, f3, f4):
    total = pd.concat([f1, f2, f3, f4], axis=0)
    total = total.drop_duplicates()
    total.reset_index(inplace=True, drop=True)
    return total

#concatenate gk files
gk = concatenate_files(gk_1, gk_2, gk_3, gk_4)
#concatenate player files
pl = concatenate_files(df_1, df_2, df_3, df_4)
#concatenate lineup files
lineups = concatenate_files(l_1, l_2, l_3, l_4)

#slice string to get matchweek and position
pl['matchweek'] = pl['matchweek'].str.slice(start=-3, stop=-1).astype(float)
gk['matchweek'] = gk['matchweek'].str.slice(start=-3, stop=-1).astype(float)
pl['position'] = pl['position'].str.slice(start=0, stop=2)
gk

lineups['matchweek'] = lineups['matchweek'].str.slice(start=-3, stop=-1).astype(float)

#map position to category
gk['position'] = 'GK'
def get_category(df):
    df.loc[df['position'] == 'GK', 'category'] = 'Goalkeeper'
    df.loc[df['position'] == 'DF', 'category'] = 'Defender'
    df.loc[df['position'] == 'RB', 'category'] = 'Defender'
    df.loc[df['position'] == 'LB', 'category'] = 'Defender'
    df.loc[df['position'] == 'CB', 'category'] = 'Defender'
    df.loc[df['position'] == 'WB', 'category'] = 'Defender'
    df.loc[df['position'] == 'FB', 'category'] = 'Defender'
    df.loc[df['position'] == 'MF', 'category'] = 'Midfield'
    df.loc[df['position'] == 'DM', 'category'] = 'Midfield'
    df.loc[df['position'] == 'CM', 'category'] = 'Midfield'
    df.loc[df['position'] == 'AM', 'category'] = 'Midfield'
    df.loc[df['position'] == 'LM', 'category'] = 'Midfield'
    df.loc[df['position'] == 'RM', 'category'] = 'Midfield'
    df.loc[df['position'] == 'WM', 'category'] = 'Midfield'
    df.loc[df['position'] == 'FW', 'category'] = 'Forward'
    df.loc[df['position'] == 'RW', 'category'] = 'Forward'
    df.loc[df['position'] == 'LW', 'category'] = 'Forward'
    return df


pl = get_category(pl)

#export lineups
def create_csv(df):
    df = df.drop_duplicates()
    df.reset_index(inplace=True, drop=True)
    df.to_csv('PL2021_lineups.csv', index=False)
    return df
#create_csv(lineups)


#export goalkeeper data
def create_csv(df):
    df = df.drop_duplicates()
    df.reset_index(inplace=True, drop=True)
    df.to_csv('PL2021_keepers.csv', index=False)
    return df

#create_csv(gk)

#export player data
def create_csv(df):
    df = df.drop_duplicates()
    df.reset_index(inplace=True, drop=True)
    df.to_csv('PL2021_players.csv', index=False)
    return df


#create_csv(pl)
