==> 04_Calculate_Player_Ratings.ipynb <==
#notebook for calculating player ratings
#import necessary packages
import pandas as pd

#import player data
PL2019_keepers = pd.read_csv('Data/2019-20/Players/PL2019_keepers.csv')
PL2019_players = pd.read_csv('Data/2019-20/Players/PL2019_players.csv')
PL2020_keepers = pd.read_csv('Data/2020-21/Players/PL2020_keepers.csv')
PL2020_players = pd.read_csv('Data/2020-21/Players/PL2020_players.csv')
PL2021_keepers = pd.read_csv('Data/2021-22/Players/PL2021_keepers.csv')
PL2021_players = pd.read_csv('Data/2021-22/Players/PL2021_players.csv')

#Goalkeeper Rating
#apply weights from Excel Solver
def calculate_gk_rating(df):
    df['Rating'] = ((df['gk_psxg'] - df['gk_goals_against'])*8.066670338)+(df['gk_save_pct']/100*2.556904605) + (df['gk_goals_against']*-5.915922982)  +(df['gk_saves']*8.969594364) + (df['gk_pct_passes_launched']/100*7.810097531) + (df['gk_psxg']*0.985661108) + (df['gk_def_actions_outside_pen_area']*1.522476572)+ (df['gk_passes']*0.228727318)

    OldMax = df['Rating'].max()
    OldMin = df['Rating'].min()
    NewMax = 100
    NewMin = 0

    OldRange = (OldMax - OldMin)
    NewRange = (NewMax - NewMin)
    df['Rating'] = (((df['Rating'] - OldMin) * NewRange) / OldRange) + NewMin
    return df

#calculate ratings for all seasons
PL2019_keepers = calculate_gk_rating(PL2019_keepers)
PL2020_keepers = calculate_gk_rating(PL2020_keepers)
PL2021_keepers = calculate_gk_rating(PL2021_keepers)

#split df in dfs for each player group
def split_player_groups(df):
    Defenders = df.drop(df[df.category != 'Defender'].index)
    Midfielders = df.drop(df[df.category != 'Midfield'].index)
    Forwards = df.drop(df[df.category != 'Forward'].index)
    return Defenders, Midfielders, Forwards

Defenders_19, Midfielders_19, Forwards_19 = split_player_groups(PL2019_players)
Defenders_20, Midfielders_20, Forwards_20 = split_player_groups(PL2020_players)
Defenders_21, Midfielders_21, Forwards_21 = split_player_groups(PL2021_players)


#Defender Rating
#apply weights from Excel Solver
def calculate_df_rating(df):
    df['Rating'] = (df['tackles']*6.141939015)+(df['progressive_carries']*9.4230058) + (df['progressive_passes']*1) + (df['assists']*7.661351766)+ (df['passes']*1.013737739) + (df['interceptions']*1) + (df['blocks']*1.526323093)  +(df['pressures']*1)  + (df['goals']*4.536477348)   + (df['passes_pct']/100*7.693134296)


    OldMax = df['Rating'].max()
    OldMin = df['Rating'].min()
    NewMax = 100
    NewMin = 0


    OldRange = (OldMax - OldMin)
    NewRange = (NewMax - NewMin)
    df['Rating'] = (((df['Rating'] - OldMin) * NewRange) / OldRange) + NewMin
    return df

#calculate ratings for all seasons
Defenders_19 = calculate_df_rating(Defenders_19)
Defenders_20 = calculate_df_rating(Defenders_20)
Defenders_21 = calculate_df_rating(Defenders_21)

#display best players over a season
df21 = Defenders_20.groupby(['player']).sum()['Rating']
df21_t = Defenders_20.groupby(['Team']).sum()['Rating']
defender_strength = df21.sort_values(ascending=False)
defender_strength_t = df21_t.sort_values(ascending=False)
defender_strength

#Midfield Rating
#apply weights from Excel Solver
def calculate_mf_rating(df):
    df['Rating'] = (df['assists']*9.896093302)+(df['goals']*8.24545131) + (df['progressive_passes']*1) + (df['progressive_carries']*1) + (df['gca']*9.974429519)  +(df['pressures']*1)  + (df['sca']*10)   + (df['dribbles_completed']*1.558838612)
    #df['Rating'] = (df['passes_pct'])

    OldMax = df['Rating'].max()
    OldMin = df['Rating'].min()
    NewMax = 100
    NewMin = 0


    OldRange = (OldMax - OldMin)
    NewRange = (NewMax - NewMin)
    df['Rating'] = (((df['Rating'] - OldMin) * NewRange) / OldRange) + NewMin
    return df

#calculate ratings for all seasons
Midfielders_19 = calculate_mf_rating(Midfielders_19)
Midfielders_20 = calculate_mf_rating(Midfielders_20)
Midfielders_21 = calculate_mf_rating(Midfielders_21)

#Forward Rating
#apply weight from Excel Solver
def calculate_att_rating(df):
    df['Rating'] = (df['goals']*5.701406026)+(df['assists']*1.185545635) + ((df['goals'] - df['xg'])*5.509198531) + ((df['assists'] - df['xa'])*6.255697541) + (df['gca']*4.072354183) + (df['sca']*2.413883824) + (df['shots_on_target']*5.056112447)
    #df['Rating'] = (df['dribbles'])

    OldMax = df['Rating'].max()
    OldMin = df['Rating'].min()
    NewMax = 100
    NewMin = 0


    OldRange = (OldMax - OldMin)
    NewRange = (NewMax - NewMin)
    df['Rating'] = ((((df['Rating'] - OldMin) * NewRange) / OldRange) + NewMin)
    return df

#calculate ratings for all seasons
Forwards_19 = calculate_att_rating(Forwards_19)
Forwards_20 = calculate_att_rating(Forwards_20)
Forwards_21 = calculate_att_rating(Forwards_21)

#merge positions to one df per season
total_19 = pd.concat([Defenders_19, Midfielders_19, Forwards_19, PL2019_keepers], axis=0)
total_20 = pd.concat([Defenders_20, Midfielders_20, Forwards_20, PL2020_keepers], axis=0)
total_21 = pd.concat([Defenders_21, Midfielders_21, Forwards_21, PL2021_keepers], axis=0)
total_19.reset_index(inplace=True, drop=True)
total_20.reset_index(inplace=True, drop=True)
total_21.reset_index(inplace=True, drop=True)
PL2019_ratings = total_19[['player', 'position', 'age', 'minutes', 'Team', 'matchweek', 'category', 'Rating', 'date']]
PL2020_ratings = total_20[['player', 'position', 'age', 'minutes', 'Team', 'matchweek', 'category', 'Rating', 'date']]
PL2021_ratings = total_21[['player', 'position', 'age', 'minutes', 'Team', 'matchweek', 'category', 'Rating', 'date']]
PL2019_ratings = PL2019_ratings.fillna('Goalkeeper')
PL2020_ratings = PL2020_ratings.fillna('Goalkeeper')
PL2021_ratings = PL2021_ratings.fillna('Goalkeeper')
PL2020_ratings
