==> 05_Generate_Player_Features.ipynb <==
#notebook to match different name spellings and generate features for players

#import necessary packages
import pandas as pd
from fuzzywuzzy import fuzz, process

#import data
#season 2019_20
PL2019_ratings = pd.read_csv('Data/2019-20/Players/PL2019_ratings.csv')
PL2019_fl = pd.read_csv('Data/2019-20/Players/FL_19.csv')
PL19_lineups = pd.read_csv('Data/2019-20/Players/PL2019_lineups.csv')
PL2019_fl.rename(columns={"round": "matchweek"}, inplace=True)
PL2019_fl['player'] = PL2019_fl['name']
#season 2020_21
PL2020_ratings = pd.read_csv('Data/2020-21/Players/PL2020_ratings.csv')
PL2020_fl = pd.read_csv('Data/2020-21/Players/FL_20.csv')
PL20_lineups = pd.read_csv('Data/2020-21/Players/PL2020_lineups.csv')
PL2020_fl.rename(columns={"round": "matchweek"}, inplace=True)
PL2020_fl['player'] = PL2020_fl['name']
#season 2020_21
PL2021_ratings = pd.read_csv('Data/2021-22/Players/PL2021_ratings.csv')
PL2021_fl = pd.read_csv('Data/2021-22/Players/FL_21.csv')
PL21_lineups = pd.read_csv('Data/2021-22/Players/PL2021_lineups.csv')
PL2021_fl.rename(columns={"round": "matchweek"}, inplace=True)
PL2021_fl['player'] = PL2021_fl['name']
PL2021_fl


#rename team names
PL2020_fl.replace(['Leeds', 'Leicester', 'Man City', 'Man Utd', 'Newcastle', 'Norwich','Spurs'], ['Leeds United', 'Leicester City','Manchester City', 'Manchester Utd', 'Newcastle Utd', 'Norwich City' ,'Tottenham'], inplace=True)
PL2021_fl.replace(['Leeds', 'Leicester', 'Man City', 'Man Utd', 'Newcastle', 'Norwich','Spurs'], ['Leeds United', 'Leicester City','Manchester City', 'Manchester Utd', 'Newcastle Utd', 'Norwich City' ,'Tottenham'], inplace=True)


#get name and team from df player ratings
players_rating = PL2021_ratings[['player', 'Team']]
players_rating = players_rating.drop_duplicates()
players_rating

#get name and team from df fantasy league
players_fl = PL2021_fl[['player', 'name', 'team']]
players_fl = players_fl.drop_duplicates()
players_fl

#get name and team from df lineups
player_lineup = PL21_lineups[['player', 'team']]
player_lineup= player_lineup.drop_duplicates()
player_lineup['name_lineup'] = player_lineup['player']
player_lineup

#merge data
players_total = pd.merge(players_fl,player_lineup,on='player', how='left')
players_total = players_total.sort_values(by='name_lineup')

#remove blanks
player_lineup['player'] = player_lineup['player'].str.strip()
players_fl['player'] = players_fl['player'].str.strip()

#perform fuzzy merge
def fuzzy_merge(df_1, df_2, key1, key2, threshold=90, limit=2):
    """
    :param df_1: the left table to join
    :param df_2: the right table to join
    :param key1: key column of the left table
    :param key2: key column of the right table
    :param threshold: how close the matches should be to return a match, based on Levenshtein distance
    :param limit: the amount of matches that will get returned, these are sorted high to low
    :return: dataframe with boths keys and matches
    """
    s = df_2[key2].tolist()

    m = df_1[key1].apply(lambda x: process.extract(x, s, limit=limit))
    df_1['matches'] = m

    m2 = df_1['matches'].apply(lambda x: ', '.join([i[0] for i in x if i[1] >= threshold]))
    df_1['matches'] = m2

    return df_1

players_merged = fuzzy_merge(player_lineup, players_fl ,'player', 'player', threshold=90, limit=1)
players_merged

#sort values to find names which could not be matched
players_merged= players_merged.sort_values(by='matches')
players_merged.reset_index(inplace=True, drop=True)
players_merged.rename(columns={"matches": "name"}, inplace=True)
players_merged

#merge df with matches with fl df to find all unmatched players and manually match them in excel sheet

players_merged['player'] = players_merged['player'].str.strip()
players_fl['player'] = players_fl['player'].str.strip()

out = pd.merge(players_merged,players_fl , on='name', how='left')

out = out[['player_x', 'team_x', 'name', 'team_y']]
out.rename(columns={"player_x": "player_rating", "team_x": "team_rating", "name": "player_fl", "team_y": "team_fl" }, inplace=True)
out

#export to excel
#out.to_excel('names_2021.xlsx')

#import manually added names
name_matching = pd.read_excel('Data/2021-22/Players/name_matching_2021.xlsx')
name_matching

#transform data for merge
name_matching = name_matching[['player_rating', 'player_fl']]
name_matching = name_matching.drop_duplicates()
name_matching.rename(columns={"player_fl": "player"}, inplace=True)

#get date out of kickoff time
PL2021_fl ['date'] = PL2021_fl['kickoff_time'].str.slice(start=0, stop=10)

#merge data
PL2021_fl['player'] = PL2021_fl['player'].str.strip()
name_matching['player'] = name_matching['player'].str.strip()
PL2021_fl_matched = pd.merge(PL2021_fl, name_matching , on='player', how='left')
#generate ID to map each time a player played a game to a specific
PL2021_fl_matched['ID'] = PL2021_fl_matched['player_rating'] + '_' + PL2021_fl_matched['date'].astype(str)

#generate ID
PL2021_ratings['ID'] = PL2021_ratings['player'] + '_' + PL2021_ratings['date'].astype(str)

#merge data on id, to get player rating for every game
PL2021_matched = pd.merge(PL2021_fl_matched, PL2021_ratings , on='ID', how='left')

#replace na (If a player was not in the squad for the match day, he will not receive a rating)
PL2021_matched['Rating'] = PL2021_matched['Rating'].fillna(0)
PL2021_matched

#rename column names
PL2021_total = PL2021_matched[['player_x','player_rating','date_x', 'team' ,'Team' ,'category', 'minutes_x' ,'minutes_y','Rating', 'bonus', 'bps', 'xP' ,'ict_index', 'influence', 'creativity','threat','total_points', 'transfers_balance', 'transfers_in', 'transfers_out', 'value', 'was_home', 'opponent_team', 'selected' ]]
PL2021_total.rename(columns={"player_x": "player_fl","player_rating": "player" ,"date_x": "date", "Team": "team_rating","team":"team",  "position_x": "position_fl", "minutes_x": "minutes_fl","minutes_y": "minutes_rating" ,"Rating": "rating"}, inplace=True)
PL2021_total


#generate game and player ID
PL2021_total['player_ID'] = PL2021_total['player'].astype(str) + PL2021_total['team_rating'] + PL2021_total['date'].astype(str)
PL2021_total['game_ID'] = PL2021_total['team_rating'] + PL2021_total['date'].astype(str)

PL2021_total

#export player dataframe including fantasy league values and calculated ratings
#PL2021_total.to_csv('Data/2021-22/Players/PL2021_total.csv')
